refactor(vue): drain trigger dialogs into useTriggerDialog composable
StopLossDialog.vue (440) and TakeProfitDialog.vue (428) were near-duplicate
script-heavy SFCs. Extract the shared form/quote/submit logic into a single
position- and mode-parameterized composable (useTriggerDialog), leaving each
SFC a verbatim template over the returned bindings (both now 119 total LOC,
well under the 400 floor). The stop-loss/take-profit residue — trigger
direction, price-comparison sign, validation i18n keys, and the getPercent
zero-guards — is selected by the `mode` argument so behaviour is preserved
exactly; the characterization tests pass unchanged. Increment of #185.
